[Custom images] Allow users to specify disk size
When building custom images, a disk is generated and attached
to the VM used to perform the build.
As more components are installed, the space on the attached disk may
run out: This patch addresses such problem by allowing users to
specify a variable disk size as an extra parameter for the script.