Increase and allow to modify timer before vm shutdown
Once the custom initialization script is executed, the
VM instance will wait 20 seconds before shutting
itself down.

If the custom initialization produces a considerable
amount of text on stdout, the instance may not have the
time to flush all of its output and Daisy may miss to
capture the signal on serial port marking success/failure.

We can solve this by simply increase the timeout before
shutdown of the VM instance (now 300 seconds).

To make it easier to modify this value in the future,
I also added an extra parameter to the
generate_custom_image.py script.