Allow to setup additional metadata for builder VM
Once you start to have multiple and complex script for custom images
it becomes very handy to be able to set additional metadata
in the VM that will run the custom init script.

As an example:
You want to run as a first step a script that set up some
common repositories and configuration and then delegates
execution to another script used to build the actual image.
You will have to run the common script first and then you
can programatically delegate the execution by using VM
metadata, which you can setup when invoking the generate_custom_image.py
script.